vault: restore original design, minor tweak to hints
The bug that affected Firefox on Windows and Linux has been fixed.
Bring back the original design, including the benefit of the "held"
object together with the dino.

While at it, perform a minor tweak to one of the hints and shorten
the maxLength a bit closer for two of the answers.

Ref <https://bugzilla.mozilla.org/show_bug.cgi?id=1761685>.